Extract Pidfile setup and deletion to a different concern
Also, rename SupervisorTest to ForkSupervisorTest, as that's the one
being tested there.
